requirements.txt
# Flask web framework
Flask==2.3.3
Flask-CORS==4.0.0

# Database (sqlite3 is part of Python standard library)

# HTTP requests
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Date/time handling
python-dateutil==2.8.2

# Async support
asyncio

# Cryptography for API signatures
cryptography==41.0.4

# Testing
pytest==7.4.2
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Development
black==23.7.0
flake8==6.0.0